+++ pages/signup-confirm/index.tsx
@@ -1,1 +1,1 @@
-export { SignUpConfirmedPage as default } from '@/pages/signUpConfirmed/SignUpConfirmedPage'
+export { SignUpConfirmedPage as default } from '@/pages/signUpConfirmed'
