+++ src/pages/signUpConfirmed/index.ts
